update register (default value to role id)

+++ app/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ProductCreateRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\HttpFoundation\Response;
 use Illuminate\Support\Str;
@@ -19,6 +20,7 @@ class ProductController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view', 'products');
         $products = Product::paginate(9);
 
         return ProductResource::collection($products);
@@ -42,8 +44,7 @@ class ProductController extends Controller
      */
     public function store(ProductCreateRequest $request)
     {
-
-
+        Gate::authorize('edit', 'products');
         $product = Product::create($request->only('title', 'title_ar', 'description', 'description_ar', 'image', 'price'));
 
         return response($product, Response::HTTP_CREATED);
@@ -80,6 +81,7 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'products');
         $product = Product::find($id);
         $product->update($request->only('title', 'title_ar', 'description', 'description_ar', 'image', 'price'));
 
@@ -94,6 +96,7 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('edit', 'products');
         Product::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
